Restore tests

+++ src/EditorFeatures/CSharpTest/CodeActions/ExtractMethod/ExtractLocalFunctionTests.cs
@@ -523,7 +523,7 @@ class Program
     static void Main()
     {
         byte z = 0;
-        Goo({|Rename:NewMethod|}(), y => 0, z, z);
+        Goo({|Rename:NewMethod|}(), y => (byte)0, z, z);
 
         static Func<byte, byte> NewMethod()
         {
@@ -561,7 +561,7 @@ class Program
     static void Main()
     {
         byte z = 0;
-        Goo({|Rename:NewMethod|}(), y => { return 0; }, z, z);
+        Goo({|Rename:NewMethod|}(), y => { return (byte)0; }, z, z);
 
         static Func<byte, byte> NewMethod()
         {
